backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
asyncpg
passlib[bcrypt]
python-jose[cryptography]
boto3
python-multipart
requests
httpx
psycopg2-binary
moviepy==1.0.3
opencv-python-headless
numpy
python-dotenv
email-validator
websockets
